[UPDATE] several changes + refators
